# Conflicts:
#   src/main/java/com/nomz/doctorstudy/common/exception/GlobalExceptionHandler.java
#   src/main/java/com/nomz/doctorstudy/config/S3Config.java
#   src/main/java/com/nomz/doctorstudy/config/security/SecurityConfig.java
#   src/main/java/com/nomz/doctorstudy/image/controller/ImageController.java
#   src/main/java/com/nomz/doctorstudy/image/exception/ImageErrorCode.java
#   src/main/java/com/nomz/doctorstudy/image/exception/ImageExceptionHandler.java
#   src/main/java/com/nomz/doctorstudy/image/request/ImageUploadRequest.java
#   src/main/java/com/nomz/doctorstudy/member/controller/AuthController.java

+++ src/main/java/com/nomz/doctorstudy/conference/service/ConferenceServiceImpl.java
@@ -7,6 +7,7 @@ import com.nomz.doctorstudy.conference.ConferenceErrorCode;
 import com.nomz.doctorstudy.conference.dto.ConferenceSearchFilter;
 import com.nomz.doctorstudy.conference.entity.ConferenceMember;
 import com.nomz.doctorstudy.conference.entity.ConferenceMemberInvite;
+import com.nomz.doctorstudy.conference.entity.ConferenceMemberInviteId;
 import com.nomz.doctorstudy.conference.repository.ConferenceMemberInviteRepository;
 import com.nomz.doctorstudy.conference.repository.ConferenceMemberRepository;
 import com.nomz.doctorstudy.conference.repository.ConferenceQueryRepository;
@@ -15,9 +16,9 @@ import com.nomz.doctorstudy.conference.request.CreateConferenceRequest;
 import com.nomz.doctorstudy.conference.request.GetConferenceListRequest;
 import com.nomz.doctorstudy.conference.request.InviteMemberConferenceRequest;
 import com.nomz.doctorstudy.conference.request.JoinConferenceRequest;
-import com.nomz.doctorstudy.conference.response.*;
 import com.nomz.doctorstudy.conference.room.RoomService;
 import com.nomz.doctorstudy.member.entity.Member;
+import com.nomz.doctorstudy.member.exception.member.MemberErrorCode;
 import com.nomz.doctorstudy.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -42,64 +43,87 @@ public class ConferenceServiceImpl implements ConferenceService {
     private final ConcurrentHashMap<Long, ReentrantLock> joinLockMap = new ConcurrentHashMap<>();
 
     @Override
-    public CreateConferenceResponse createConference(CreateConferenceRequest request) {
+    public Long createConference(Member requester, CreateConferenceRequest request) {
         Conference conference = Conference.builder()
                 .title(request.getTitle())
+                .host(requester)
                 .memberCapacity(request.getMemberCapacity())
+                .isFinished(false)
                 .build();
         conferenceRepository.save(conference);
 
         log.info("[new conference] id={}, title={}", conference.getId(), conference.getTitle());
 
-        return CreateConferenceResponse.builder()
-                .conferenceId(conference.getId())
-                .build();
+        return conference.getId();
     }
 
     @Override
     @Transactional
-    public GetConferenceResponse getConference(Long conferenceId) {
-        Conference conference = conferenceRepository.findById(conferenceId)
+    public Conference getConference(Long conferenceId) {
+        return conferenceRepository.findById(conferenceId)
                 .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
-
-        return GetConferenceResponse.builder()
-                .id(conference.getId())
-                .title(conference.getTitle())
-                .memberCapacity(conference.getMemberCapacity())
-                .build();
     }
 
     @Override
-    public List<GetConferenceListResponse> getConferenceList(GetConferenceListRequest request) {
-        List<Conference> conferenceList = conferenceQueryRepository.getConferenceList(
+    public List<Conference> getConferenceList(GetConferenceListRequest request) {
+        return conferenceQueryRepository.getConferenceList(
                 ConferenceSearchFilter.builder()
                         .title(request.getTitle())
-                        .memberCapacity(request.getMemberCapacity())
+                        .isFinished(request.getIsFinished())
                         .build()
         );
+    }
 
-        return conferenceList.stream()
-                .map(GetConferenceListResponse::of)
+    @Override
+    public List<Member> getConferenceParticipantList(Long conferenceId) {
+        List<ConferenceMember> conferenceMembers = conferenceMemberRepository.findByConferenceId(conferenceId);
+
+        return conferenceMembers.stream()
+                .map(ConferenceMember::getMember)
                 .toList();
     }
 
     @Override
-    public void startConference(Long conferenceId) {
+    public void openConference(Long conferenceId) {
+        Conference conference = conferenceRepository.findById(conferenceId)
+                .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
+
+        if (joinLockMap.containsKey(conferenceId)) {
+            throw new BusinessException(ConferenceErrorCode.CONFERENCE_ALREADY_STARTED);
+        }
         joinLockMap.put(conferenceId, new ReentrantLock());
         roomService.createRoom(conferenceId);
     }
 
     @Override
+    public void startConference(Long conferenceId) {
+        Conference conference = conferenceRepository.findById(conferenceId)
+                .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
+
+
+    }
+
+    @Override
     public void finishConference(Long conferenceId) {
+        Conference conference = conferenceRepository.findById(conferenceId)
+                .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
+
+        if (!joinLockMap.containsKey(conferenceId)) {
+            throw new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_IN_PROCESS);
+        }
+
         roomService.removeRoom(conferenceId);
         joinLockMap.remove(conferenceId);
     }
 
     @Override
     @Transactional
-    public JoinConferenceResponse joinConference(Member requester, Long conferenceId, JoinConferenceRequest request) {
-        List<String> peerIds = null;
+    public List<String> joinConference(/*Member requester, */Long conferenceId, JoinConferenceRequest request) {
+        List<String> peerIds;
         ReentrantLock lock = joinLockMap.get(conferenceId);
+        if (lock == null) {
+            throw new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_IN_PROCESS);
+        }
 
         lock.lock();
         try {
@@ -112,42 +136,39 @@ public class ConferenceServiceImpl implements ConferenceService {
         Conference conference = conferenceRepository.findById(conferenceId)
                 .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
 
+        /*
         ConferenceMember conferenceMember = ConferenceMember.builder()
                 .id(new ConferenceMember.ConferenceMemberId(conferenceId, requester.getId()))
                 .conference(conference)
-                .member(requester)
+                //.member(requester)
                 .build();
         conferenceMemberRepository.save(conferenceMember);
+        */
 
-        return JoinConferenceResponse.builder()
-                .existingPeerIds(peerIds)
-                .build();
+        return peerIds;
     }
 
     @Override
     @Transactional
-    public InviteMemberConferenceResponse inviteMemberConference(Member requester, Long conferenceId, InviteMemberConferenceRequest request) {
+    public void inviteMemberConference(Member requester, Long conferenceId, InviteMemberConferenceRequest request) {
         Conference conference = conferenceRepository.findById(conferenceId)
                 .orElseThrow(() -> new BusinessException(ConferenceErrorCode.CONFERENCE_NOT_FOUND_ERROR));
 
         if (requester.getId() != conference.getHost().getId()) {
             // TODO: MemberErrorCode로 변경
-            throw new BusinessException(CommonErrorCode.FORBIDDEN);
+            throw new BusinessException(CommonErrorCode.FORBIDDEN, "호스트 유저만이 초대할 수 있습니다.");
         }
 
         // TODO: Member 에러코드 변경
         Long inviteeId = request.getInviteeId();
         Member member = memberRepository.findById(inviteeId)
-                .orElseThrow(() -> new BusinessException(CommonErrorCode.BAD_REQUEST));
+                .orElseThrow(() -> new BusinessException(CommonErrorCode.BAD_REQUEST, "초대할 멤버의 아이디를 찾을 수 없습니다."));
 
         ConferenceMemberInvite conferenceMemberInvite = ConferenceMemberInvite.builder()
-                .id(new ConferenceMemberInvite.ConferenceMemberInviteId(conferenceId, inviteeId))
+                .id(new ConferenceMemberInviteId(conferenceId, inviteeId))
                 .conference(conference)
                 .member(member)
                 .build();
         conferenceMemberInviteRepository.save(conferenceMemberInvite);
-
-        return InviteMemberConferenceResponse.builder()
-                .build();
     }
 }
